+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 [unreleased diff]
 
+## [2.1.0] - 2024-08-28
+
+[2.1.0 diff]
+
 ### Added
 
 - Convar to remember the audio volume set by the player using audio controls.
@@ -133,7 +137,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Finishing message
 - Configuration file
 
-[unreleased diff]: https://github.com/D4isDAVID/loadscreen/compare/v2.0.0...main
+[unreleased diff]: https://github.com/D4isDAVID/loadscreen/compare/v2.1.0...main
+[2.1.0]: https://github.com/D4isDAVID/loadscreen/releases/tag/v2.1.0
+[2.1.0 diff]: https://github.com/D4isDAVID/loadscreen/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/D4isDAVID/loadscreen/releases/tag/v2.0.0
 [2.0.0 diff]: https://github.com/D4isDAVID/loadscreen/compare/v1.2.0...v2.0.0
 [1.2.0]: https://github.com/D4isDAVID/loadscreen/releases/tag/v1.2.0